Autodetect netcat binaries
Some distributions such as archlinux ship with the openbsd-nc as the
default /usr/bin/nc binary, which does not support the -U option.
Often, an alternative binary "netcat" is available on this systems which
represents the gnu-netcat variant.
This patch autodetects the "netcat" binary and uses this over "nc" where
available